Analyze Falsing on Expanded Notif Shade.
Ensure that gestures on the expanded notification shade are picked up
properly by the FalsingManager. This adds `SHADE_DRAG` and
`QS_COLLAPSE` falsing interaction types to cover these cases.

Also, try to limit calling on the FalsingManager until a gesture is
actually ready to occur, instead of querying it while deciding what
the current gesture is.

Fixes: 184042853
Test: manual
Change-Id: I1f651614d4d1520bac1a2375319f69fb1017505f