Ensure teardown() only triggered on the right NetworkAgent.
This change ensures that teardown() is only ever triggered when the
NetworkAgent matches the one that the unwanted() callback is triggered
on. This ensures that in cases where the unwanted() is called as a
result of unregister(), or even after a new NetworkAgent is created (and
stored in the mNetworkAgent field), the teardown does not occur.

Bug: 184304972
Test: atest FrameworksVcnTests
Change-Id: Ic8ab66ff1e48d43be784974cda7574d0191e557f