Wait OCCLUDE transition until screen turning on activity becomes ready.
When the screen is off, WindowManger usually skips running an app
transition animation when updating visibility of activities. However, in
case the activity is going to turn on the screen, we should not skip an
app transition animation even if the screen is not yet on.

I extended ActivityRecord#okToAnimate, so that it can ignore screen
status and it can be used with ActivityRecord#canTurnScreenOn predicate.

Bug: 175687166
Bug: 182975035
Change-Id: Id8079df2027becadc9c5e8266149f6eb76b467b6
Test: Receive phone call while the screen is off.