Change remaining trusted layers from BQL to Blast layer
All BQL have now have an underlying BSL that will render the content. This causes
issues with trusted layers since the input info is not propagated down
from the top layer to the BSL that contains the buffer. This causes
touch issues since a non trusted layer is blocking another layer.

Since trusted layers can only be created by the system, convert all BQL
that are trusted layers to be BSL so we can specificy the input directly
on the layer that contains the buffer.

Test: Accessibility no longer blocks input
Bug: 185973018
Change-Id: Ia208c570f2a694fb65c73ef0bec0ab34b0bcc8a0