Set background color to the AppBarLayout
During the page transition, the back arrow on the action bar is acting
as a shared element. The transition seems to render the shared element
first and then the other parts of the UI. As the background color of the
AppBarLayout is transparent, so during the rendering process, the UI
will flash/blink for a few milliseconds and the content behind the
AppBarLayout will be shown.

Hence, set a background color to the AppBarLayout to fix this issue.

Fixes: 186701900
Test: rebuild and review Display Settings page transition
Change-Id: Id009b47d983adafd8b92af585b5089c68de0afc3