Remove @GuardedBy on a final, immutable field
Remove @GuardedBy on a final, immutable field as it is causing
errorprone warnings on a work-in-progress commit.

ContentProviderHolder.mUri is accessed in multiple places without
holding mLock, and it is final and documented as immutable so should be
safe to access without mLock. This commit should reduce ErrorProne
warnings now and in future.

Bug: 185786624
Test: build / treehugger only
Change-Id: Ic7fbe550ff6f538f336cfb859f66daa94844ca28