[Device Controls] Fix background not showing up.
Root cause was that `this` on lines 353 and 354 was referring to the
ValueAnimator, not the clipLayer.getDrawable().apply{} that was started
on line 330. I refactored the code a bit so it uses fewer Kotlin idioms
but should be less error-prone, but let me know if you'd rather I just
update lines 353 and 354 to specify which `this` it's referring to.

Test: manual
Fixes: 187903455

Change-Id: I3505c1fc80f7dda6d6cc318b0dfa2754171b8a2b