Temporarily increase sched policy for dozing
If always-on-display is enabled, there will be an animation for
transiting to AOD screen. To prevent the animating process from
using lower process scheduling group because of wakefulness,
set it as animating to gain top-app group with a short duration.
So the animation may be smoother.

Though AOD may be disabled or other reasons that there won't be
an animation, it should be fine by always setting the state.
Because the target process just keeps idle, and the duration
is short.

Bug: 188253927
Test: Press power key to turn off screen, check:
      adb shell dumpsys activity o | grep " PERS.*systemui"
      It will show "T/ /PERU" =
          SCHED_GROUP_TOP_APP
          PROCESS_STATE_PERSISTENT_UI
      After a few seconds it will show "R/ /BFGS" =
          SCHED_GROUP_RESTRICTED
          PROCESS_STATE_BOUND_FOREGROUND_SERVICE

Change-Id: Ia331a84cf70d28fdbfd4b9b02629c29539197fd5